+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 # [1.0.0](https://github.com/jhipster/generator-jhipster-nodejs/tree/v1.0.0)
 
+-   Check controller generator with lint and tests [issue #92](https://github.com/jhipster/generator-jhipster-nodejs/issues/92)
+
 # [1.0.0-beta.2](https://github.com/jhipster/generator-jhipster-nodejs/tree/v1.0.0-beta.2)
 
 -   Exclude server TS files from client lint and test increasing quality on the generated [issue #83](https://github.com/jhipster/generator-jhipster-nodejs/issues/83)